implement exercises 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -36,6 +36,19 @@ blogsRouter.post('', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+        return response.status(401).json({error: 'token missing or invalid'})
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(404).end()
+    } else if (blog.user.toString() !== decodedToken.id) {
+        return response.status(401).end()
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
